break out analytic components
