+++ client/components/index.js
@@ -28,7 +28,7 @@ export {default as AdminTeams } from './AdminTeams'
 export {default as AdminCompanies} from './AdminCompanies'
 export {default as AdminUserModal} from './AdminUserModal'
 export {default as AddUserModal} from './AddUserModal'
-export {default as UserCreatedProjects} from './Analytics/user_created_projects'
+export {default as UserActiveProjects} from './Analytics/user_active_projects'
 export {default as UserProjectTypes } from './Analytics/user_project_types'
 export {default as CompaniesHomePage} from './Company/companies_home_page'
 export {default as UserCompletedTasks} from './Analytics/user_completed_tasks'
