+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "An application to streamline projects/assignments for Treasury Sales Managers, Officers, and Analysts within Financial Institutions.",
   "engines": {
-    "node": ">= 7.0.0"
+    "node": "9.2.1"
   },
   "main": "index.js",
   "scripts": {
@@ -29,8 +29,6 @@
     "chart.js": "^2.8.0",
     "compression": "^1.7.4",
     "connect-session-sequelize": "^4.1.0",
-    "d3": "^5.9.2",
-    "draft-js": "^0.10.5",
     "express": "^4.17.0",
     "express-flash": "0.0.2",
     "express-session": "^1.16.1",
@@ -46,7 +44,6 @@
     "pg": "^6.1.2",
     "pg-hstore": "^2.3.2",
     "prop-types": "^15.7.2",
-    "quill": "^1.3.6",
     "radium": "^0.21.2",
     "react": "^16.8.6",
     "react-animations": "^1.0.0",
@@ -57,7 +54,6 @@
     "react-redux": "^5.1.1",
     "react-router-dom": "^4.3.1",
     "react-select": "^2.4.3",
-    "react-vis": "^1.11.7",
     "redux": "^3.6.0",
     "redux-logger": "^2.8.1",
     "redux-thunk": "^2.3.0",
